requirements.txt
hydra-core
mediapipe
pytorch-lightning
sentence-transformers
torchcodec~=0.5
zipfile-inflate64
einops-exts
torchview
mne
tokenizers
safetensors
scipy~=1.15.2
requests~=2.32.3
dotenv~=0.9.9
python-dotenv~=1.1.0
numpy~=2.2.4
torch~=2.7.0
h5py~=3.13.0
transformers~=4.51.3
einops~=0.8.1
torchinfo~=1.8.0
tqdm~=4.67.1
scikit-learn~=1.6.1
pandas~=2.2.3
huggingface-hub~=0.30.2
torchaudio~=2.7.0
opencv-python~=4.11.0.86
torchvision~=0.22.0
librosa~=0.11.0
matplotlib~=3.10.3
omegaconf~=2.3.0
ipywidgets~=8.1.7
python-dateutil~=2.9.0.post0
psutil~=7.0.0
pytz~=2025.2
moviepy~=2.2.1
ffmpeg-python~=0.2.0
SpeechRecognition~=3.14.3
accelerate
lightning